Allow the site to be installed locally
To make it easier for developers to contribute to the project, I updated README.md and CONTRIBUTING.md with installation and contribution instructions.

I also added a Gemfile with the "github-pages" gem, which is the recommended way of installing Jekyll and making sure all the dependency versions are in sync with GitHub Pages.

Read more about the gem here: https://github.com/github/pages-gem